File::getSelectLines: Don't apply open interval adjustment when selection is a single line.
The trailing point of the selection is not included, so getSelectLines
applies an adjustment. If the both points are on different lines to code
correctly identifies the trailing point.
But if but points are on the same line and the point the code checks is
at the start of the line, it wrongly applies the adjustment.
It is safe to never adjust when both points are in the same line, because
either both points are equal and at the start of the line, or the trailing
point can't be at the start of the line.