TextCursor,File: Move cursor and normal selection into TextCursor class and quick and dirty rewrite block selection to use completely independent variables.
This work is done in a large commit, because there didn't seem to be a good
way to split this without massively more work.

The block selection and multi insert features are still very much work in
progress. The previous implementation sharing the variables from the
normal cursor turned out to be problematic because for block selection
code unit in line based positioning has many limitation. So use the
forced rework to switch to use column based addressing for the block
selection features.
This forces usage of special activateBlockSelection and
disableBlockSelection whenever switching between selection modes,
but at least it's easy to see what happens when these are explicit.

Missing:
- adjust scroll position does not work yet.
- Highlighting outside of the line text does not work yet.