File: Use signals emitted on changes to our cursor and our line markers to ensure the scroll position is correct and send cursor position signals.
Changes to the documents from cursors outside of this class can move or
scroll position line marker. If that happens we need to ensure the scroll
position still leaves the current cursor position visible.

On the other hand changes to the documents from cursors outside of this
class can also move the cursor (or what is used for scrolling purposes
in block select mode). In this case we also need to ensure the scroll
position still leaves the changed cursor position visible.

Before this commit the cursor position signal have been send
unconditionally from the code in adjustScrollPosition() as that was a
code path called on every cursor change. Now that we have an actual
notification of cursor position changes from the document classes switch
to using that to avoid sending spurious change notifications when
recalculating the scroll position for other reasons.

The tests for text searching currently use the cursor position signals
and have been adjusted to work with the new somewhat changes amount of
emission of these.