File: Save and restore y scroll position when inserting text skip automatic LineMarker adjustment.
When inserting a line break at the start of the top most line of the view
the scroll position line marker will be adjusted, but when inserting in
the same view, the scroll position shouldn't change and the new line
should be visible.

So save and restore the scroll position around all inserts that might add
a new line.