Document,DocumentSnapshot: Add per line userData pointer.
This allows users of the document to attach auxilary data or caches.

Uses a shared_ptr to manage lifetime because snapshots and undo history
need to keep these objects alive.

If the snapshot is used in a different thread, it is the users
responsibility to avoid race conditions when accessing the user data.