Document,DocumentSnapshot: Add a document revision counter.
This allows detecting changes to the document by simply comparing a number.

The revision only increases (until wrap around). This is true even for an
undo.

Capture the revision of the document at time of the snapshot creation in
DocumentSnapshot to detect if work done on a snapshot still matches the
document state.

Also share the documents live revision with the snapshot to allow querying
if the snapshot still is current, allowing to abandon work that would only
be useful if the revision of the snapshot matches the documents revision at
completion time.