Document: Add support for multi line regex search and work around searching with invalid utf18.
Invalid utf16 is not supported by QRegularExpression (which uses libpcre2).
It either checks that the string is valid utf16 or if that is disabled,
invalid utf16 is documented to cause undefined behavior.

To work around that replace anything that is not allowed with U+FFFD.
That way we at least enable searching valid parts of lines that contain
invalid utf16. As we checked for validity anyway, disable the now redundant
check in libprce2.